requirements.txt
PyYaml
argparse
jinja2
wget
